requirements.txt
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-telegram-bot>=20.7
streamlit>=1.32.0
vectorbt>=0.26.1
plotly>=5.18.0
